refactor(constants): src 폴더로 이동
