+++ src/components/main/Modal.ts
@@ -1,7 +1,7 @@
 import type { Restaurant } from '../../type';
 import Component from '../Component';
 import restaurantsStorage from '../../lib/restaurantsStorage';
-import { CATEGORIES } from '../../utils/constants';
+import { CATEGORIES } from '../../constants';
 
 interface ModalProps {
   toggleModal: () => void;
