+++ src/components/main/restaurant/RestaurantList.ts
@@ -1,7 +1,7 @@
 import type { Category, Restaurant, SortBy } from '../../../type';
 import Component from '../../Component';
 import RestaurantListItem from './RestaurantListItem';
-import { DEFAULT_CATEGORY } from '../../../utils/constants';
+import { DEFAULT_CATEGORY } from '../../../constants';
 
 interface RestaurantListProps {
   currentCategory: Category;
